% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM.plotMetrics.r
\name{CM.plotMetrics}
\alias{CM.plotMetrics}
\title{Plot changes of the banks (erosion and aggradation)}
\usage{
CM.plotMetrics(cmgo.obj, set = "set1", cl = NULL, d = NULL)
}
\arguments{
\item{cmgo.obj}{the global object of type list containing data and parameters created with \code{\link[=CM.ini]{CM.ini()}}}

\item{set}{the reference data set}

\item{cl}{the range of centerline points to be plotted, if NULL (default) the full channel length will be plotted, if
a vector of two elements is provided (e.g. c(200, 500)) this cl range is plotted, if a string is provided (e.g. "cl1"),
the range defined in \code{par$plot.cl.ranges$cl1} will be plotted}

\item{d}{the distance range of the centerline downstream to be plotted, NULL (default) cl defintions are taken, if a
single value (e.g. d=500) is given 50 m around this distance is plotted, if a vector with two elements is given (e.g.
c(280, 620)) this distance range will be plotted}
}
\value{
desc
}
\description{
If multiple channel surveys are present (time series analyses) and the reference centerline mode is
used, this function allows to plot the shift of the banks (bank erosion and aggradation) along the channel.
}
\details{
\code{CM.plotMetrics()} allows to plot the position of banks with regard to the centerline in downstream
direction. If only one survey is present, the plot will only show the distance of the right and left bank
to the centerline over the centerline distance (downstream length of the channel). If multiple surveys of a
channel exists (time series analyses) the differences of the channel banks can be calculated if a reference
centerline is used (reference centerline mode). The differences (bank ersosion and aggradation) are plotted
with this function.

Without passing something to the function, the banks of the whole channel are plotted. To specify a region
use the \code{cl} and \code{d} parameters. See the parameter definition and the example section for details.
}
\examples{

# open demo
cmgo.obj = CM.ini("demo3")

# example 1: plot the distance of channel banks to centerline
CM.plotMetrics(cmgo.obj)

# example 2: plot the change of the channel banks (aggradation/erosion)
CM.plotMetrics(cmgo.obj, set="set2")

# example 3: plot only a range
CM.plotMetrics(cmgo.obj, set="set2", cl=c(800, 850))
CM.plotPlanView(cmgo.obj,  set="set2", cl=c(800, 850)) # compare with plan view map

}
\author{
Antonius Golly
}
