% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM.plotWidth.r
\name{CM.plotWidth}
\alias{CM.plotWidth}
\title{Plot channel width}
\usage{
CM.plotWidth(cmgo.obj, set = "set1", set.compare = NULL, cl = NULL, d = NULL)
}
\arguments{
\item{cmgo.obj}{the global object of type list containing data and parameters created with \code{\link[=CM.ini]{CM.ini()}}}

\item{set}{the primary data set to plot ("set1" by default)}

\item{set.compare}{the secondary data set to plot (optional)}

\item{cl}{the range of centerline points to be plotted, if NULL (default) the full channel length will be plotted, if
a vector of two elements is provided (e.g. c(200, 500)) this cl range is plotted, if a string is provided (e.g. "cl1"),
the range defined in \code{par$plot.cl.ranges$cl1} will be plotted}

\item{d}{the distance range of the centerline downstream to be plotted, NULL (default) cl defintions are taken, if a
single value (e.g. d=500) is given 50m around this distance is plotted, if a vector with two elements is given (e.g.
c(280, 620)) this distance range will be plotted}
}
\value{
desc
}
\description{
Plot the channel width of the whole channel (default) or for a portion (use \code{cl} argument). The function can
also compare two data sets if they have the same reference centerline.
}
\details{
If more than one data set is defined in the global data object, of

Details
}
\examples{

# open demo
cmgo.obj = CM.ini("demo2")

# example 1: plot channel width of whole channel
CM.plotWidth(cmgo.obj)

# example 2: plot channel width of a defined range (centerline points)
CM.plotWidth(cmgo.obj, cl=c(800, 1000))

# example 2: plot channel width of a defined range (distance downstream)
CM.plotWidth(cmgo.obj, d=c(200, 600))

}
\author{
Antonius Golly
}
