% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM.reduceCenterline.r
\name{CM.reduceCenterline}
\alias{CM.reduceCenterline}
\title{Reduce centerline points}
\usage{
CM.reduceCenterline(object, set = NULL)
}
\arguments{
\item{object}{the global object of type list containing data and parameters created with \code{\link[=CM.ini]{CM.ini()}}}

\item{set}{an optional argument for processing a specific data set, if \code{NULL} all available data sets are used}
}
\value{
the global data object
}
\description{
Create an equally spaced centerline by decreasing its spatial resolution while preserving its length.
}
\details{
The spatial resolution of the centerline depends on the spatial resolution of the polygon generated from the bank points. Thus,
it is directly dependent from the parameter par$polygon.bank.interpolate.max.dist. If you have chosen a value smaller
than the channel width, the resulting resolution of the centerline can be very high. This results in a large number of
centerline points, which are not necessary and introduce high computational costs during further calculation. With
the function CM.reduceCenterline you can again decrease the resolution to a given value.
}
\examples{

# get the demo data set
cmgo.obj = CM.ini("demo2")

# reduce centerline resolution
cmgo.obj = CM.reduceCenterline(cmgo.obj)

}
\author{
Antonius Golly
}
