% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM.resampleCenterline.r
\name{CM.resampleCenterline}
\alias{CM.resampleCenterline}
\title{Resample centerline points}
\usage{
CM.resampleCenterline(cmgo.obj, set = NULL)
}
\arguments{
\item{cmgo.obj}{the global object of type list containing data and parameters created with \code{\link[=CM.ini]{CM.ini()}}}

\item{set}{an optional argument for processing a specific data set, if \code{NULL} all available data sets are used}
}
\value{
the global data object
}
\description{
Create an equally spaced centerline with a given interval length.
}
\details{
The spatial resolution of the centerline depends on the spatial resolution of the bank points which is
directly dependent on the parameter par$polygon.bank.interpolate.max.dist. A small interpolation distance is necessary for complex
bed shapes. However, for further metric analyes a large number of centerline points is not necessary. Since they can introduce
high computational costs during further calculation you can resample the centerline intervals with this function CM.resampleCenterline().
decreasing the resolution to a given value. This will have slight impact on the length of the centerline. Since you are losing
detail a coarser interval will decrease the length of the centerline.
}
\examples{

# get the demo data set
cmgo.obj = CM.ini("demo2")

# resample centerline resolution
cmgo.obj = CM.resampleCenterline(cmgo.obj)

}
\author{
Antonius Golly
}
