% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM.writeData.r
\name{CM.writeData}
\alias{CM.writeData}
\title{Write data}
\usage{
CM.writeData(object)
}
\arguments{
\item{object}{the global object of type list containing data and parameters created with \code{\link[=CM.ini]{CM.ini()}}}
}
\value{
This file function does not return any value.
}
\description{
Write output files with channel metrics and the current workspace.
}
\details{
\code{CM.writeData()} allows you to write the results to output files and to an R workspace file. 
The outputs written depend on the settings in the parameter object. If \code{cmgo.obj$par$workspace.write = TRUE} 
(default is FALSE) a workspace file is written containing the global data object. The 
filename is defined in \code{cmgo.obj$par$workspace.filename}. Further, ASCII tables can be written 
containing the centerline geometry and the calculated metrics. If \code{cmgo.obj$par$output.write = TRUE} 
(default is FALSE) an output file for each data set is written to the output folder specified 
in \code{cmgo.obj$par$output.dir}. The file names are the same as the input filenames with the 
prefixes cl_* and metrics_*. All parameters regarding the output generation are listed in \code{\link[=CM.par]{CM.par()}}.
}
\examples{
# get demo data (find instructions on how to use own data in the documentation of CM.ini())
cmgo.obj = CM.ini("demo2")

# example 1: write workspace
cmgo.obj$par$workspace.write = TRUE
CM.writeData(cmgo.obj)

# example 2: write output files
cmgo.obj$par$output.write = TRUE
cmgo.obj$par$output.write.centerline = TRUE
cmgo.obj$par$output.write.metrics = TRUE
cmgo.obj$par$output.dir   = "custom_output_folder"
cmgo.obj$par$workspace.write = FALSE

CM.writeData(cmgo.obj)

}
\author{
Antonius Golly
}
