% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.r
\docType{data}
\name{demo}
\alias{demo}
\title{Raw input demo data set to run the full stack of cmgo functions.
Contains a list with sublists $par and $data.}
\format{An object of class \code{list} of length 2.}
\usage{
demo
}
\description{
Raw input demo data set to run the full stack of cmgo functions.
Contains a list with sublists $par and $data.
}
\keyword{datasets}
