% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo2.r
\docType{data}
\name{demo2}
\alias{demo2}
\title{Fully processed demo data set to run test the output and plotting functions of cmgo.
Contains a list with sublists $par and $data.}
\format{An object of class \code{list} of length 2.}
\usage{
demo2
}
\description{
Fully processed demo data set to run test the output and plotting functions of cmgo.
Contains a list with sublists $par and $data.
}
\keyword{datasets}
