% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo3.r
\docType{data}
\name{demo3}
\alias{demo3}
\title{Fully processed demo data set to run test the output and plotting functions of cmgo.
In contrast to demo data set 2 this data set uses the reference centerline mode.
Contains a list with sublists $par and $data.}
\format{
An object of class \code{list} of length 2.
}
\usage{
demo3
}
\description{
Fully processed demo data set to run test the output and plotting functions of cmgo.
In contrast to demo data set 2 this data set uses the reference centerline mode.
Contains a list with sublists $par and $data.
}
\keyword{datasets}
